[MIG] account_invoice → move, new fields name and buttons, remove intercompany, FIX groups for Validate buttons
