+++ fermente_account_invoice_workflow/models/__init__.py
@@ -1,1 +1,1 @@
-from . import account_invoice
+from . import account_move
