+++ fermente_account_invoice_workflow/readme/CONTRIBUTORS.md
@@ -1,2 +1,3 @@
-- Sylvain LE GAL (<https://www.twitter.com/legalsylvain>)
+- Sylvain LE GAL
 - Julien WESTE
+- Quentin DUPONT
